Add support for configurable server listen port
This patch adds support for configuring the UDP port that the DHCP
helper will listen on when processing responses from the DHCP server.
The port value is provided from the json config file.

If the listen port is anything other than the standard port 67, we
create another UDP socket to handle incoming traffic on the configured
listen port.  The original UDP socket on port 67 will be used only to
send traffic.

This ensures that the source port is correct in DHCP packets relayed by
the relay agent.  All traffic from the relay in any direction, towards
clients or server, will now have the UDP source port 67.

Signed-off-by: Jonas Holmberg <jonas.holmberg@westermo.se>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>